+++ Server/Program.cs
@@ -13,28 +13,9 @@ namespace YourNote.Server
     public class Program
     {
 
-        private static string[] connectionData = GetConnectionData();
-        private static string Host = connectionData[0];
-        private static string Port = connectionData[1];
-        private static string DBname = connectionData[2];
-        private static string User = connectionData[3];
-        private static string Password = connectionData[4];
-
-        private static readonly string connectionString = $" Host={Host}; Port={Port}; Database={DBname};" +
-                $" Username={User}; Password={Password};";
-
+     
         public static void Main(string[] args)
         {
-
-            var serviceProvider = CreateServices();
-
-            // Put the database update into a scope to ensure
-            // that all resources will be disposed.
-            using (var scope = serviceProvider.CreateScope())
-            {
-                UpdateDatabase(scope.ServiceProvider);
-            }
-
 
 
 
@@ -42,34 +23,7 @@ namespace YourNote.Server
         }
 
 
-        private static IServiceProvider CreateServices()
-        {
-            return new ServiceCollection()
-                // Add common FluentMigrator services
-                .AddFluentMigratorCore()
-                .ConfigureRunner(rb => rb
-                    // Add SQLite support to FluentMigrator
-                    .AddPostgres()
-                    // Set the connection string
-                    .WithGlobalConnectionString(connectionString)
-                    // Define the assembly containing the migrations
-                    .ScanIn(typeof(AddLogTable).Assembly).For.Migrations())
-                // Enable logging to console in the FluentMigrator way
-                .AddLogging(lb => lb.AddFluentMigratorConsole())
-                // Build the service provider
-                .BuildServiceProvider(false);
-        }
-
-
-        private static void UpdateDatabase(IServiceProvider serviceProvider)
-        {
-            // Instantiate the runner
-            var runner = serviceProvider.GetRequiredService<IMigrationRunner>();
-
-            // Execute the migrations
-            runner.MigrateUp();
-        }
-
+       
 
         public static IWebHost BuildWebHost(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
@@ -79,8 +33,7 @@ namespace YourNote.Server
                 .UseStartup<Startup>()
                 .Build();
 
-        private static string[] GetConnectionData() => Environment.GetEnvironmentVariable("PGPASSDATA").Split(':');
-
+        
 
 
     }
