+++ Shared/Models/MigrationOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace YourNote.Shared.Models
+{
+    public class MigrationOptions
+    {
+        public string ConnectionString { get; set; }
+
+        public long? Version { get; set; }
+    }
+}
